move & enum APP:STARTUP
