+++ app/state/actions/index.js
@@ -1,0 +1,1 @@
+export const APP_STARTUP = 'APP:STARTUP'
